# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaign-routes:
    get:
      operationId: listCampaignRoutes
      tags: [campaign-routes]
      summary: List Rocket Engine campaign routes in the current organization
      parameters:
        - name: campaignId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of campaign routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignRoute"
    post:
      operationId: createCampaignRoute
      tags: [campaign-routes]
      summary: Create a campaign route (owner/admin/manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRouteInput"
      responses:
        "201":
          description: Created campaign route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRoute"
  /organizations/{organizationId}/campaigns/{campaignId}/rocket-setup:
    put:
      operationId: configureRocketCampaign
      tags: [campaign-engine]
      summary: Configure a campaign's multi-number, multi-template Rocket setup
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/CampaignIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RocketCampaignSetupInput"
      responses:
        "200":
          description: Rocket campaign setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RocketCampaignSetup"
...
  /campaign-routes/{routeId}:
    patch:
      operationId: updateCampaignRoute
      tags: [campaign-routes]
      summary: Update a campaign route (owner/admin/manager only)
      parameters:
        - $ref: "#/components/parameters/RouteIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRouteUpdate"
      responses:
        "200":
          description: Updated campaign route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRoute"
    delete:
      operationId: deleteCampaignRoute
      tags: [campaign-routes]
      summary: Delete a campaign route (owner/admin/manager only)
      parameters:
        - $ref: "#/components/parameters/RouteIdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: integer
...
    CampaignIdParam:
      name: campaignId
      in: path
      required: true
      schema:
        type: integer
    RouteIdParam:
      name: routeId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    CampaignRoute:
      type: object
      required: [id, campaignId, campaignName, phoneNumberId, phoneNumber, wabaExternalId, templateId, templateName, priority, configuredTps, currentTps, queueDepth, status, isSample, createdAt, updatedAt]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        campaignName: { type: string }
        phoneNumberId: { type: integer }
        phoneNumber: { type: string }
        wabaExternalId: { type: ["string", "null"] }
        templateId: { type: ["integer", "null"] }
        templateName: { type: ["string", "null"] }
        priority: { type: string, enum: [High, Normal, Low] }
        configuredTps: { type: integer }
        currentTps: { type: integer }
        queueDepth: { type: integer }
        status: { type: string, enum: [Active, Throttled, Error, Paused] }
        isSample: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CampaignRouteInput:
      type: object
      required: [campaignId, phoneNumberId]
      properties:
        campaignId: { type: integer }
        phoneNumberId: { type: integer }
        templateId: { type: ["integer", "null"] }
        priority: { type: string, enum: [High, Normal, Low] }
        configuredTps: { type: integer, minimum: 1 }
        currentTps: { type: integer, minimum: 0 }
        queueDepth: { type: integer, minimum: 0 }
        status: { type: string, enum: [Active, Throttled, Error, Paused] }
    CampaignRouteUpdate:
      type: object
      properties:
        campaignId: { type: integer }
        phoneNumberId: { type: integer }
        templateId: { type: ["integer", "null"] }
        priority: { type: string, enum: [High, Normal, Low] }
        configuredTps: { type: integer, minimum: 1 }
        currentTps: { type: integer, minimum: 0 }
        queueDepth: { type: integer, minimum: 0 }
        status: { type: string, enum: [Active, Throttled, Error, Paused] }
    RocketCampaignNumberInput:
      type: object
      required: [phoneNumberId, configuredTps]
      properties:
        phoneNumberId: { type: integer }
        configuredTps: { type: integer, minimum: 1 }
    RocketCampaignSetupInput:
      type: object
      required: [numbers, templateIds]
      properties:
        numbers:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/RocketCampaignNumberInput"
        templateIds:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: integer }
        priority:
          type: string
          enum: [High, Normal, Low]
          default: Normal
    RocketCampaignSetup:
      type: object
      required: [campaignId, numberCount, templateCount, aggregateTps, routes]
      properties:
        campaignId: { type: integer }
        numberCount: { type: integer }
        templateCount: { type: integer }
        aggregateTps: { type: integer }
        routes:
          type: array
          items:
            $ref: "#/components/schemas/CampaignRoute"
...