# lib/api-zod/src/generated/types/campaignRoute.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignRoutePriority } from './campaignRoutePriority';
import type { CampaignRouteStatus } from './campaignRouteStatus';

export interface CampaignRoute {
  id: number;
  campaignId: number;
  campaignName: string;
  phoneNumberId: number;
  phoneNumber: string;
  /** @nullable */
  wabaExternalId: string | null;
  /** @nullable */
  templateId: number | null;
  /** @nullable */
  templateName: string | null;
  priority: CampaignRoutePriority;
  configuredTps: number;
  currentTps: number;
  queueDepth: number;
  status: CampaignRouteStatus;
  isSample: boolean;
  createdAt: Date;
  updatedAt: Date;
}